+++ CHANGES.rst
@@ -9,6 +9,8 @@ unreleased
 Release 0.1
 -----------
 
+- Layer chooser is now a directive and has a background-image.
+
 - All clicks on the map result all data available to that location
 
 - ExtentAggregat is the default card displaying an extent summary of